src/graphRAG/graphRAG/requirements.txt
networkx
matplotlib
pandas
python-dotenv
langchain
langchain-openai
langchain-community
sentence-transformers
scikit-learn
numpy
